changed totp to match version 8
The current regex extracts the first token, which is technically TOTP 8, but Spotify doesn't actually validate the TOTP version in the URL. So, for the time being, this works without needing to verify the extracted token's TOTP version. This fix doesn't fix any breaking issues, and currently, using a TOTP token with mismatching versions doesn't break anything, but probably will in the future, so fuck future nonsense.